deployment update

+++ deployment/stopApp.sh
@@ -1,3 +1,8 @@
-#!/bin/sh
+#!/bin/bash
 
-kill `ps -ae |grep node |grep -v grep|awk '{print $1}'` | exit 0
+if [ "ps -ae |grep node |grep -v grep|awk '{print $1}'" -gt 0 ] 
+	kill `ps -ae |grep node |grep -v grep|awk '{print $1}'` | exit 0
+then
+fi
+
+
